// lib/utils/element-type.enum.ts
export enum ElementType {
  command = 'command',
  commandHandler = 'handler',
  commandSpec = 'command_spec',
  event = 'event',
  eventHandler = 'event_handler',
  eventUpdater = 'updater',
  query = 'query',
  queryHandler = 'query_handler',
  eventHandlerSpec = 'event_handler_spec',
  eventUpdaterSpec = 'event_updater_spec',
  queryHandlerSpec = 'query_handler_spec',
}